Setup sqlite3
The sqlite3 module was installed into the project. A separate JS module is responsible for the sql functionality
of the app. A basic enterExpense functoin was added along with a getExpenses to make sure that the sqlite3 module is functioning properly at the most basic level. The database file will be created in the user's 'userDataPath' directory, which is specific to the operating system.