Add form to query expenses by date range
There is another form now that allows the user to enter
two dates, a start and end date. Functionality will be added
in a later commit to query the database for all expenses in between the two dates entered.

In the css, the #ipcForm id was generalized to just apply
the text-align to all forms.