Add the ability to query expenses in date range
The user can now use a form to enter a start date and end date to query their expenses with.

If a start date is not entered, a date representing the 'start of time' will be used. Similarly, if an end date is not entered, a date representing the 'end of time' will be used.