Refactor getSqlString
Changed the name of the function to formatStringForSql as that seemed like a more appropriate name considering its purpose. Also added a line of code in this fucntion to account for single quotes that may be in the string. For sql queries, a single quote in the middle of string needs to be escaped by adding another single quote before it. A line of code was added that uses regex to replace all instances of single quotes in the string with two single quotes.